Extract common_state module from conn
