+++ rustls/src/client/tls12.rs
@@ -1,5 +1,6 @@
 use crate::check::{inappropriate_handshake_message, inappropriate_message};
-use crate::conn::{self, CommonState, ConnectionRandoms, Side, State};
+use crate::common_state::{send_cert_verify_error_alert, CommonState, Side, State};
+use crate::conn::ConnectionRandoms;
 use crate::enums::ProtocolVersion;
 use crate::enums::{AlertDescription, ContentType, HandshakeType};
 use crate::error::{Error, InvalidMessage, PeerMisbehaved};
@@ -739,7 +740,7 @@ impl State<ClientConnectionData> for ExpectServerDone {
                 &st.server_cert.ocsp_response,
                 now,
             )
-            .map_err(|err| conn::send_cert_verify_error_alert(cx.common, err))?;
+            .map_err(|err| send_cert_verify_error_alert(cx.common, err))?;
 
         // 3.
         // Build up the contents of the signed message.
@@ -765,7 +766,7 @@ impl State<ClientConnectionData> for ExpectServerDone {
             st.config
                 .verifier
                 .verify_tls12_signature(&message, &st.server_cert.cert_chain[0], sig)
-                .map_err(|err| conn::send_cert_verify_error_alert(cx.common, err))?
+                .map_err(|err| send_cert_verify_error_alert(cx.common, err))?
         };
         cx.common.peer_certificates = Some(st.server_cert.cert_chain);
